OKVL implementation first version. This version does minimal thing to get OKVL working.
Although it's incomplete in terms of integration into experiments/tests code,
I separated this as one commit because it concisely contains the "core" of OKVL implementation.

Missing things:

- Experiments/tests folder are tentatively disabled.
Needs several minor changes in those folders to get them compile with the new lock manager.

- Non-unique index and secondary index is missing.
These two types of indexes will be implemented later.
They are related to the benefits of OKVL, but mostly orthogonal in terms of implementation.

- Performance experiments for OKVL paper.
Ideally we should address the above and then do experiments with really interesting workload.
However, we should quickly move on to more interesting things, so we will do some quick "hacks"
to do necessary performance experiments without the features.

- Several other clean-up/optimizations.