Removed "returnRC" mechanism from log archiver
The threads used in the log archiver make use of a "returnRC" flag to
store any error codes encountered during execution. The idea was to
detect errors on asynchronous threads without interrupting the execution
of the main thread. However, the mechanism didn't make much sense, since
we only checked for such errors when the thread exited, and we used the
W_COERCE macro for that. This means that early detection of such errors
was not being useful in any way. Therefore, I removed the returnRC
mechanism and errors are thrown immediately using W_COERCE. Since our
project is more experimental/research prototype, we don't have to be so
worried about catching errors early on in the hope of fixing them.