Fixed RunScanner to deal with index blocks
If an index is build during run generation, its data is stored at the
end of the run file, in N blocks appended after the actual data blocks.
This commit extends RunScanner to stop scanning when the index blocks
are reached, by keeping track of the total number of data blocks. To
that end, I implemented a simple method to read metadata from the last
block of the file and, together with the file size, determine the number
of data and index blocks.