Removed device manager and its uses
Shore/Zero always had the concept of a device, which was supposed to be
composed of multiple volumes -- in the same way a linux device is
composed of multiple partitions. However, this destinction was never
really realized, and a device was always assumed to contain just one
volume.

In a refactoring which was supposed to be done many years ago, I removed
the device manager and all its uses, delegating to volume operations
where applicable. It seems like things will keep working as expected,
but more tests will be required to see if anything is broken by this
change.