Removed SM level 4
This begins a series of commits in which I'll attempt to remove the
concept of "SM levels". The goal is to have independent components (bf,
log, lock, io, etc.) which can be constructed independently, and a much
"thiner" ss_m which simply orchestrates the components and delegates API
calls to them.