New implementation of B-tree split
Existing B-tree split (three options implemented by Wey for minimal and
full logging) were not working with restore because they either had
dependencies to the buffer pool or between pages. We should strictly
follow the phisiological loggin principle that pages can be restored
independently, which is crucial for instant restore. Furthermore, full
logging is a requirement of restore in general.

See the comments on the new methods for more details.