New version of btree split with SSX
Using an SSX to implement btree split in order to avoid undo of bulk
delete on foster parent. For single-page recovery, chains of both pages
involved should lead to the same log record. Whether to redo foster
parent or child depends on the page id of the given fixable_page_h. For
restore, we will simply duplicate multi-page log records so that both
pages are restored independently.