Fixed behavior of multi-page logrec REDO
To determine which page is being replayed when calling the REDO method
of a multi-page log record, we compare the PID of the given page image
with the "pid2" field of the log record. Comparing with the main pid
field (i.e., construct_pid()) is no longer valid because of the
replication in the log archiver.