Root page is loaded on demand during restore
During restart, root pages are preloaded in the buffer pool
using bf_tree_m::install_volume during log analysis. For restore, this
approach would not work, because when the volume is re-mounted after
marking it failed, the metadata (specifically, the stnode cache) must be
restored before the root page can be determined.

The current commit loads the root page on demand at fix_root. When it
tries to access the volume metadata (with vol_t::get_store_root()), it
should block until it has been restored. After this, the root is loaded.