Added options to tweak restore behavior
- trySinglePass: if true, restore scheduler will return page IDs in
sequential order if there are no requests available, meaning that
single-pass restore is performed together with on-demand restore.
- reuseRequestBuffer: if true, read requests from the volume manager
may reuse the restore buffer and copy requested page contents into the
target frame in the buffer pool after restore. This avoids an additional
read of a page which was just recently on main memory.

These two behaviors used to be on by default