Removed MAX_VOL_COUNT and fixed bf_cleaner volume mgmt
All places that ask for the maximum number of mounted volumes now access
vol_m::MAX_VOLS instead of the constant MAX_VOL_COUNT. However, this
behavior is not entirely correct, since we need to distinguish between
the maximum number of mounted volumes (MAX_VOLS) and the maximum number
of volumes that can ever be created (MAX_VID). Right now, the two are
undistinguishable, but we may want to fix that in the future.

Simplified volume management in the buffer cleaner. Cleaner thread is
assigned to volume based on a simple mod operation and reserved slot 0
for "no volume" was eliminated from certain arrays, since it was not
used (at least one cleaner thread must always exist).