New (logged) sx_add_backup operation for volumes
A volume now contains a list of FS paths to be used as backup files.
These can be added with the vol_m::sx_add_backup method. The operation
is part of persistent system state and thus it must be logged,
checkpointed, and recovered. Logging and recovery are done,
chekpointing will come next.