Fixed B-tree split by deleting body space
B-tree items consist of a head, which is a fixed-length header stored in
the head array from left to right, and a body, which can be
variable-length (but multiple of sizeof(item_body)) and is stored right
to left.

The previous split operation only deleted entries from the head array,
without properly deleting the associated bodies.