Several bug fixes on eager log archiver
- Reader thread cannot run eagerly, because endLSN must be in sync with
log archiver. This is because log consumer expects a skip log record on
the endLSN position, and it must be artificially inserted by reader.
- Reader activation fixed to use a single signal instead of a loop.
Thanks to proper lock acquisition, we guarantee that the signal is
always caught.
- Reader correctly fetches partial blocks across activation cycles,
i.e., if then endLSN is in the middle of a block, following read will
fetch data into the rest of the block, so that scanner continues from
the same offset where it left off.
- LogArchiver keeps track of the endLSN of the last activation, to avoid
being activated on a region which was already covered before. This is
important for cases when the archiver runs faster than the log grows.
- Shutdown of log archiver is synchronous now, i.e., it sends
synchronous shutdown to all sub-components and joins on all threads.
- Writer thread forked from the beginning rather than on demand.
Eliminated writerForked flag and fixed shutdown.
- Small fix on heap workspace deallocation (array delete was wrong)
- Added log archiver shutdown to SM destructor.