Fixed archive index building and loading
- Changed building mechanism to support multiple runs being indexed
while writer thread only writes out one at a time. The previous
assumption that the last run entry on the index was always the one being
(1) appended to by BlockAssembly and (2) written out by WriterThread was
completely wrong, since both threads run independently without
synchronization. New mechanism makes use of a "lastFinished" array
index, which points to the last entry already serialized by the writer
thread. All entries up to this one can be probed -- the remaining ones
were created by BlockAssembly but not serialized yet. (Fixes #5)

- Removed filter mechanism of RunInfo objects. The goal was to eliminate
runs that are known to not contain a given PID from a probe. However, the
strategy proved not very effective in practice, and no one was
working on it. Without filter information, RunInfo is much leaner and
easy to serialize/deserialize.

- Fixed serialization and deserialization of the archive index,
following the new design.