Restore buffer reusing defaults to false
- When a page request arrives during restore, it goes into a request
queue together with its buffer pool frame address. When the page is
then finally restored, we try to reuse the restore workspace and copy
the contents to the given frame, thus avoiding an extra I/O to read the
restored page. However, this is not really working yet, so I am turning
it off by default.
- Extra checks and debug messages.