Fixed bugs on log archiver consumer
Previous mechanism was not working when the (fake) skip log record
spanned two neighboring blocks. Changed the whole mechanism to not rely
on fake skip log records but solely on nextLSN, which is what I should
have implemented from the start.

- Reader does not insert fake skip; it simply stops readin when the
endLSN is contained in the block just read (or if it's the first byte of
the next block).
- LogScanner does not treat skip log record in any special way. It is up
to the caller to see that it got a skip and stop invoking the scanner on
the same block.
- Also fixed nasty little bug when a log record ignored by the log
scanner fits exactly at the end of a block (such that toSkip ==
remaining in the code)