Implemented mechanism to read only whole blocks in archiver
- When the option to read only whole blocks is set, the log archiver
only activates if there is at least one block worth of log to be read.
Furthermore, it adjusts the endLSN parameter passed to the consumer so
that it is always at a block boundary, such that only whole blocks are
read instead of partial ones.
- This eliminates the assumption that endLSN is always a valid log
record. Thus, the consumer and scanner had to be adapted to handle that
case. The scanner now returns a length parameter, which helps the
consumer decide whether it should request a new block or not.
- However, there are cases where the log is either growing too slow or
it has stopped altogether. To avoid the log archiver getting stuck
because it only reads whole blocks, a wait mechanism was implemented. If
the activation window is too small, the archiver waits a certain grace
period. If after that, the current durable_lsn of the log is still equal
than the current activation's endLSN, then the log is growing too slow,
and the consumer proceeds with partial blocks. This happens at least at
the end of an experiment, where transactions have finished and we are
waiting for the archiver to complete.
- Added a stat to count how often this "log too slow" situation
happens. The grace period, which is an SM option, should be increased if
it happens too often.