Implemented log truncation
- Log truncation is useful to load a database and leave behind a clean
database and a minimal amount of log. In our case, the log starts a new
partition and copies the contents from the last checkpoint into it,
adjusting the LSNs and checkpoint information. Then, all previously
existing partitions can be deleted.

- This commit also includes a bugfix on the backward log scanner, which
would fail if the checkpoint begin was the first log record of the first
existing partition. In that case, because there is no previous
partition, it could not get the "nxt" LSN field correctly.