Fixed embarassing bug on log archiver
- The end boundary of a run file was being generated as simply the LSN
of the last log record in the last block, which makes absolutely no
sense. The boundary should be the *maximum* LSN within that run. To fix
that, I changed the BlockAssembly behavior to keep track of the max LSN
in each block (instead of the last), and the WriterThread to keep track
of the maxLSN among all blocks of the current run.

(I guess the bug hasn't been detected yet because all my tests used ony
one run)