Added backup LSN
- Each backup file is now equipped with an LSN which corresponds to the
most recent update guaranteed to be reflected in its pages. It is
extracted by looking at the archive last LSN before the backup is taken.
- The purpose of this LSN is to prune runs which are not required during
restore. If the backup was taken when N runs were available, and during
restore there are M runs, at most M-N runs will be replayed for each
segment.