Added memory fences to cleaner daemon
- Master and slave threads of cleaner daemon rely on several flags to
control their execution, but no synchronization mechanism was in place
to guarantee the visibility of updates made to such flags. This means
that threads could come to a halt.
- I added fences to all places where I think such visibility is
required. The whole orchestraiton of such threads could actually be made
simpler, but I'm guessing that all cases are covered for now.