Fixed logrec duplication in log archiver
- Multi-page log records such as btree_split neeed to be duplicated in
the log archiver so that they can be reached when retoring either page
independently.
- The previous code had a bug in which the duplicated copy could end up
in a different run, meaning that its LSN would be below the low boundary
of the new run.
- To fix that, duplication is now handled directly in the ArchiverHeap
code, and we guarantee that both log records are added as an indivisible
unit to the same run.