Fixed sharp backup taking
- To make sure that a sharp backup is taken from the current tail of the
recovery log, a stronger version of log archive flushing was
implemented, in which not only we wait for all log records to be
consumed and written out to archive blocks, but also forcibly close the
currenr run and create a new one. This also makes sure that the new run
has the requested LSN as its upper boundary.