Implemented variable-bucket log archive index
- This new index is used if a parameter bucketSize > 0 is passed to the
cosntructor. If it works really well we may make it the default and
remove code exclusive to the "old" index.
- Instead of keeping entries in the index with offsets in fixed
increments of the block size and the first PID which happens to occur in
each block (i.e., just like a B-tree), this new index records the first
page ID of each logical "bucket" and whatever arbitrary offset this page
ID occurs in the file. A bucket is a set of page IDs belonging to fixed
logical ranges (e.g., page IDs 0-99, then 100-199, and so on). For use
with instant restore, a bucket should be the same size of a segment.
- The advantage of the new index is that only the amount of log required
to restore a segment is read from the log archive. In the old
implementation, segments with few updates caused reads of a whole block
(e.g., 1MB) where actually only a few bytes where used. The
multiple-segment technique implemented in an earlier commit did not work
very well for many runs. We hope this new index will achieve the goal of
better read efficiency reagrdless of the number of runs.
- In essence, this index is superior because it allows us to jump
directly to the offset of a segment.