Extended LA index code for future improvements
- Extensions below were implemented, but they don't change the behavior
of restore or archiving yet. I am hoping they will facilitate future
improvements that I'm planning.
- Opening an archive scan actually creates a list of ProbeResult objects
before constructing a heap. This will allow us to adjust the number of
segments to restore before opening any run file, based solely on index
information.
- For the same reasons as above, also extended ProbeResult to include
the offset and PID of the end of the scan, and extended the probeInRun
method accordingly.
- Binary search in findEntry now returns the index of the entry on the
entry array, instead of the offset directly.
- Field ioUnit was added to ArchiveScanner, but it's not used so far.