Removed checks for is_swizzling_enabled
- The parts of the code that deal with maintenance of the parent pointer
in a control block only work if swizzling is enabled.
- It should be enabled by default, so the check should return true,
which it isn't doing. But that's not the point...
- I don't see why we should only maintain a parent pointer if swizzling
is enabled. (Well, when you think about it, with swizzling disabled,
there is no need for eviction to do something with the parent, so indeed
the pointer is not needed. But I'm leaving it like this for now. Proper
fix should come with new eviction algorithm)