Disabled page pinning mechanism
- Pinning mechanism is supposed to allow latch-free operations on pages,
but it was basically broken. It was used only for page eviction, which
resulted in many inconsitencies and deadlocks (which is why I am
currently fixing the eviction algorithm). Therefore, I simply disabled
it for now.
- The pin count itself is still used, so that an evicted page is not
latched and used by mistake -- even though this should not be necessary
with a better design. However, it is not updated with a CAS anymore,
requiring instead a latch on the page prior to every access.
- Also removed some methods related to write-order dependency, which is
also broken.