Basis for new eviction algorithm
- Gave up on "hierarchical traversal" idea because it is simply broken.
A B-tree cannot be properly traversed without acquiring latches --
otherwise we could apply the same latch-free mechanism of eviction to
normal traversals! I was getting all kinds of inconsistencies, and tried
to fix them for a while, but the mechanism will always be unreliable.
- New approach simply sweeps through the control blocks sequentially and
tries to latch a page and its parent (in reverse order). If the latching
succeeds (done conditionally) and the page is a leaf, then it is
selected for eviction.
- In terms of locality and hit ratio, we don't expect this to work
better than LRU or clock, but it should still be better than random.
- This approach is still not really working, because parent pointer is
hard to maintain, and so we quickly come to the situation where all
pages in buffer are unavailable for replacement.
- New approach comming in the next commits!