Acquiring EX latch to read page into a frame
- The _fix_nonswizzled method always latches the frame in the mode
requested by the caller. However, if the mode is SH and the page must be
fetched from disk, two threads may attempt to read different pages
simultaneously into the same frame. Sure, it could be handled by
verifying and allowing only one read to succeed, but it is simpler and
more error prone to acquire an EX latch regardless of the requested
mode. If SH was requested, a downgrade is performed once the read is
completed and the frame is properly initialized.