Removed unused vol field from alloc_cache
- Volume was only used in a single assertion, so it was not required as
a member variable. Data is accessed indirectly through bf_fixed, which
also supports a null volume, as long as no flushes are performed (i.e.,
read-only mode). This is possible because the constructor directly takes
a file descriptor as parameter.
- This functionality was implemented to extend the dbinspect command of
zapps with allocation information. We basically need to load the
alloc_cache of aa volume without creating or mounting a vol_t object.