New archive index probe method and multiple-segment restore
- Whole probing/scanning mechanism of log archive was updated and is
much simpler now, also supporting multiple-segment restore.
- Open method now takes a start pid and the number of page ids to be
scanned (segment size), as well as a min and a max read size.
- Probe method returns a list of probe result objects, which simply
contain a start offset and an end page ID on which to stop scanning.
In order to achive reads greater than the given min size, the end pid
is adjusted based on neighboring segments, but going no further than the
max read size.
- After initial probe, open method readjusts endPid of all run scanners
based on the maximum value collected.
- Restore manager was also updated to work with the new mechanism.