Fixed log archive scan to use correct segment borders
- For the first segment which overlaps with metadata, special treatment
is required on the restore part. In a previous commit, I made it ignore
the first pages by opening the scan directly on the first data pid.
However, this broke the scan because the end of the scan (previously
given by segmentSize) was not adjusted accordingly.
- Other than adjusting both the begin and end boundaries, these must be
independent from the "segment size" parameter of the scan, which is used
only to expand the length of the scan into multiple segments. To fix
that, I reintroduced the endPID argument. Now, startPid-endPid tell
where the scan must begin and finish. If multiple-segment scanning is
desired, then the segmentSize parameter tells which valid values can be
assumed by endPID.