New log archive file format
- Log records are not stored contiguously in runs, with a skip log
record determining the EOF. I.e., the same format as log files. The only
difference is that index blocks are stored at the end.
- The old format kept log record in blocks in the same format as
produced by BlockAssembly. However, the format is not beneficial for
scans on the log archive, since unless opening a file at the exact
beginning of a block, I/O sizes will be smaller than a block. The block
format was maintained for index blocks, however.
- The log archive generation methods (mostly in the WriterThread) had to
be changed. For RunScanner, a LogScanner object is required to parse the
continuous stream of log records.