New metadata management (stnode and alloc caches) [big commit]
- Allocation is now based on extents of 32k pages, where the first page
contains the bitmap for that extent. New extents are simply appended, so
that DB expands dynamically without a fixed quota. This also means that
restart and restore don't have to treat metadata pages differently,
since allocation log records apply to the alloc_page of the
corresponding extent.
- Extents can be assigned to stores, so that pages of a store are
clustered. For that end, stnode page was exteded so that it keeps track
of the last extent allocated to each store. Store number 0 is reserved
for extents which are not assigned to any store.
- Both stnode and alloc caches now work in a decoupled way, i.e., they
only manage internal data structures and generate log records on the
corresponding page IDs. This means that the actual metadata page contents
are not managed directly, and pages are not written directly, but
updated by the decoupled cleaner.
- Since decoupled checkpoint and cleaner were not merged in from Lucas'
branch yet, the code is currently not working for shutdown and restart,
i.e., it only works for newly loaded datasets.
- Removed BackupManager and bf_fixed
- Removed methods related to flushing buffer/log until a certain LSN
- Cleaned up SM methods related to disk info, quota, etc.
- Removed the "StoresToFree" and all related stuff from xct_t. This
seems to be legacy code. Current stnode cache does not support deletion
of stores anyway.
- Removed all stuff related to store_operation. Any operation related to
stores should be its own system transaction in stnode_cache_t.
- Removed volume header since metadata is covered by the first (alloc)
page of each extent and the stnode page with PID 1.
- Enforcing only one mounted volume at a time with an assertion. In the
near furute, we'll eliminate support for multiple logical volumes.