Clean-up of the log analysis mechanism [part 1]
- Removed forward scan
- Moved common logic of marking a page dirty, a transaction active,
etc., into the chkpt_t class, making the code more modular and with less
repetition.
- Log analysis now has a more generic treatment of log records, without
a lot of specific tasks executed depending on the logrec type. Actions
such as acquiring a lock, marking a page dirty, and a transaction active
do not depend on the logrec type -- they are executed for any logrec
that represents an update.
- Removed a lot of complexities and unnecessary optimizations (e.g., the
"mapCLR" stuff)