Implemented new log read buffer for log_core::fetch
- Buffers a certain number of log partitions, starting from the latest
one backwards.
- Asynchronous thread loads buffer in chunks of 32MB in reverse
chronological order.
- log_core keeps track of LSN range already buffered and get log record
from buffers in case of a hit.
- Since a big sequential LSN range is what is buffered, a pair of
variables suffices and there is no need to track individual log pages
with a hash table. Given the pattern of log reads during restart
recovery, this mechanism is also quite effective.
- Also moved log analysis log record inside init method, so that we
also count the buffer initialization as startup time in the experiments.