Added best-effort partition recycling mechanism
- Support for sm_log_max_partitions options, which limits the number of
partition files in the log directory. At each partition creation, old
partitions are deleted proactively. If the limit is reached, a forced
deletion is attempted, but it only works if the min active LSN collected
on the last checkpoint allow it -- hence the best-effort approach.
- A true recycling mechanism is only possible with log reservations,
which we removed because they were messy and not fully effective
(required "fudge factors").