Fixed some B-tree bugs
- 1) Steal operations when splitting, growing, etc. must always use the
"opaque" version of child pointers, otherwise a swizzled pointer will be
implicitly unswizzled without setting the _swizzled flag on the child,
leading to an inconsistent state.
- 2) When moving records to a new foster child during split, first
branch key was actually copied twice: once in pid0 and once in the first
element slot. This results in a duplicated child pointer, which works ok
(or rather, the error is masked) until we activate swizzling.
- These bugfixes should finally fix swizzling and eviction. They cost me
quite a few hours of hair-pulling frustration.